[DEV-155] refactor: 프로필 카드 버튼 컴포넌트와 프로필 카드 모달 컴포넌트 분리
